241114.2100
Minor tweaks, fixes and refactoring. Dropped support for Windows 8.1 and below.
